+++ crates/objc2/src/top_level_traits.rs
@@ -1,4 +1,4 @@
-use alloc::ffi::CString;
+use core::ffi::CStr;
 use core::ptr::NonNull;
 
 use crate::__macros::define_class::get_initialized_ivar_ptr;
@@ -153,7 +153,7 @@ pub unsafe trait Message: RefEncode {
 /// let cls = MyObject::class();
 ///
 /// // We can now access properties of the class.
-/// assert_eq!(cls.name().to_str().unwrap(), MyObject::NAME);
+/// assert_eq!(cls.name(), MyObject::NAME);
 ///
 /// // And we can send messages to the class.
 /// //
@@ -236,11 +236,7 @@ pub unsafe trait ClassType: Message {
     /// The name of the Objective-C class that this type represents.
     ///
     /// `T::NAME` is the `const` version of `T::class().name()`.
-    ///
-    /// This must not contain any NUL bytes.
-    //
-    // TODO: Convert this to CStr next time we do big changes to ClassType.
-    const NAME: &'static str;
+    const NAME: &'static CStr;
 
     /// Get a reference to the Objective-C class that this type represents.
     ///
@@ -443,7 +439,7 @@ pub type Ivars<T> = <T as DefinedClass>::Ivars;
 /// use objc2::runtime::NSObjectProtocol;
 /// // Get a protocol object representing the `NSObject` protocol
 /// let protocol = <dyn NSObjectProtocol>::protocol().expect("NSObject to have a protocol");
-/// assert_eq!(<dyn NSObjectProtocol>::NAME, protocol.name().to_str().unwrap());
+/// assert_eq!(<dyn NSObjectProtocol>::NAME, protocol.name());
 /// # // Ensure Foundation links on GNUStep
 /// # let _cls = objc2::class!(NSObject);
 /// ```
@@ -461,11 +457,7 @@ pub type Ivars<T> = <T as DefinedClass>::Ivars;
 /// ```
 pub unsafe trait ProtocolType {
     /// The name of the Objective-C protocol that this type represents.
-    ///
-    /// This must not contain any NUL bytes.
-    //
-    // TODO: Convert this to CStr next time we do big changes to ProtocolType.
-    const NAME: &'static str;
+    const NAME: &'static CStr;
 
     /// Get a reference to the Objective-C protocol object that this type
     /// represents.
@@ -485,17 +477,11 @@ pub unsafe trait ProtocolType {
     /// protocol, e.g. if the program is not properly linked to the framework
     /// that defines the protocol.
     fn protocol() -> Option<&'static AnyProtocol> {
-        get_protocol(Self::NAME)
+        AnyProtocol::get(Self::NAME)
     }
 
     #[doc(hidden)]
     const __INNER: ();
-}
-
-// Outlined to reduce code size
-fn get_protocol(name: &str) -> Option<&'static AnyProtocol> {
-    let name = CString::new(name).expect("protocol name must be UTF-8");
-    AnyProtocol::get(&name)
 }
 
 // Split into separate traits for better diagnostics